chore: Conver settings to components
